server/requirements.txt
Flask==2.3.3
requests==2.31.0
google-genai
flask-cors
python-dotenv==1.0.0
twilio==8.10.0